+++ KFXAdditions/Classes/UILabel+KFXAdditions.m
@@ -1,0 +1,18 @@
+
+
+#import "UILabel+KFXAdditions.h"
+
+@implementation UILabel (KFXAdditions)
+
+-(BOOL)kfx_setText:(NSString*)text withDefault:(NSString*)defaultText{
+
+    if (text == nil || [text isEqual:[NSNull null]] || text.length == 0) {
+        self.text = defaultText;
+        return NO;
+    }else{
+        self.text = text;
+        return YES;
+    }
+}
+
+@end